Adiciona BasePerson e move arquivos de validação para nova pasta
