+++ src/main/groovy/icaruswings/utils/validations/ValidateEmail.groovy
@@ -1,4 +1,4 @@
-package icaruswings.utils
+package icaruswings.utils.validations
 
 import java.util.regex.Matcher
 import java.util.regex.Pattern
